character sheet web rework started
