+++ Client/lib/src/features/admin/common/presentation/admin_menu_screen.dart
@@ -10,6 +10,7 @@ import 'admin_menu_item.dart';
 import 'manage_group_screen.dart';
 
 class AdminMenuScreen extends StatelessWidget {
+  static const String name = "Admin";
   static const String route = "/admin";
 
   const AdminMenuScreen({super.key});
