+++ Client/lib/src/features/characters/presentation/desktop/character_sheet.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:kalinar/src/features/characters/domain/character.dart';
+
+class CharacterSheet extends StatelessWidget {
+  final Character character;
+
+  const CharacterSheet(this.character, {super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(child: Text(character.name));
+  }
+}
